Move the stack package from pkg to model
