+++ model/stack/main.go
