+++ pkg/apps/installer_test.go
@@ -14,13 +14,13 @@ import (
 
 	"github.com/cozy/afero"
 	"github.com/cozy/checkup"
+	"github.com/cozy/cozy-stack/model/stack"
 	"github.com/cozy/cozy-stack/pkg/apps"
 	"github.com/cozy/cozy-stack/pkg/apps/appfs"
 	"github.com/cozy/cozy-stack/pkg/config/config"
 	"github.com/cozy/cozy-stack/pkg/consts"
 	"github.com/cozy/cozy-stack/pkg/couchdb"
 	"github.com/cozy/cozy-stack/pkg/prefixer"
-	"github.com/cozy/cozy-stack/pkg/stack"
 )
 
 var localGitCmd *exec.Cmd
